Split web progress snapshot helpers
