+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/progress/snapshots/ProgressReviewScheduleOverlay.kt
@@ -69,8 +69,8 @@ internal fun validateProgressReviewScheduleCacheTimeZone(
 
 // Decide whether the local review-schedule fallback can replace the server-base schedule.
 //
-// Gating contract (must match iOS FlashcardsStore+ProgressSnapshot.swift:155-184 and
-// web apps/web/src/appData/progress/snapshots/progressSnapshots.ts:175-185):
+// Gating contract (must match iOS ProgressAggregation.swift:242-253 and
+// web apps/web/src/appData/progress/snapshots/progressReviewScheduleSnapshots.ts:31-40):
 //   1. The local cache must be fully hydrated for the scope (isLocalReviewScheduleScopeHydrated).
 //      Without full hydration the local cards table cannot represent the user-wide schedule.
 //   2. localFallback.totalCards - pendingCardTotalDelta must equal serverBase.totalCards.
